requirements.txt
blinker==1.9.0
click==8.1.8
DateTime==5.5
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
names==0.3.0
packaging==24.2
pytz==2025.2
six==1.17.0
transitions==0.9.2
Werkzeug==3.1.3
zope.interface==7.2
requests
python-dotenv